// lib/types.ts
export type UserRole =
  | 'masyarakat'
  | 'mahasiswa'
  | 'perangkat_desa'
  | 'dosen'
  | 'universitas'
  | 'admin';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  is_verified: boolean;
  avatar_url?: string;
  phone?: string;
  created_at?: string;
  profile?: MahasiswaProfile | PerangkatDesaProfile | DosenProfile | UniversitasProfile;
}

export interface MahasiswaProfile {
  id: number;
  user_id: number;
  nim: string;
  jurusan: string;
  fakultas: string;
  angkatan: string;
  ktm_url?: string;
  total_jam_kkn: number;
  target_jam_kkn: number;
  kelompok_id?: number;
  kelompok?: Kelompok;
}

export interface PerangkatDesaProfile {
  id: number;
  user_id: number;
  nama_desa: string;
  kecamatan: string;
  kabupaten: string;
  provinsi: string;
  jabatan: string;
  sk_url?: string;
  latitude?: number;
  longitude?: number;
  kepala_desa?: string;
}

export interface DosenProfile {
  id: number;
  user_id: number;
  nip: string;
  fakultas: string;
  departemen: string;
  bidang_keahlian?: string;
  kuota_bimbingan: number;
  jumlah_kelompok_binaan: number;
}

export interface UniversitasProfile {
  id: number;
  user_id: number;
  kode_kampus: string;
  nama_universitas: string;
  lppm_ketua?: string;
}
// ...
export interface Kelompok {
  id: number;
  nama_kelompok: string;
  kode_kelompok: string;
  ketua_id: number;
  ketua_nama: string;
  dosen_id?: number;
  dosen_nama?: string;
  pos_kebutuhan_id?: number;
  pos_kebutuhan_judul?: string;
  desa_nama?: string;
  anggota: {
    id: number;
    user_id: number;
    nama: string;
    nim: string;
    jurusan: string;
    role_kelompok: 'Ketua' | 'Anggota';
    avatar_url?: string;
  }[];
  total_anggota: number;
  status_program: 'perencanaan' | 'pelaksanaan' | 'penyusunan_luaran' | 'selesai';
  progres_persen: number;
}